Add CRUD functionnalities on Formation Rest API

+++ webServices/FormationRestApi/src/main/java/formation/bll/FormationBll.java
@@ -10,6 +10,10 @@ public class FormationBll implements FormationDAO {
 	
 	private FormationJdbcImpl dao ;
 	
+	public FormationBll() {
+		dao = new FormationJdbcImpl();
+	}
+	
 	@Override
 	public Formation getFormationById(int id) {
 		if(id>0) {
@@ -26,39 +30,44 @@ public class FormationBll implements FormationDAO {
 
 	@Override
 	public boolean insertFormation(Formation formation) {
-		if(formation.getIdRespCat()!=0 && formation.getStheme()!=null) {
-			dao.insertFormation(formation);
-			return true;
+		
+		if(formation.getIdRespCat()!=0 && formation.getStheme()!=null && formation.getCodeForm()!=null && !dao.codeExist(formation.getCodeForm())) {
+			return dao.insertFormation(formation);
+			//return true;
 		}else {
-			System.err.println("Impossible d'ajouter cette formation car ID RespCat ou ID Theme est pas valide");
+			System.err.println("Impossible d'ajouter cette formation car ID RespCat ou ID Theme est pas valide ou alors ce Code Formation existe déja");
 			return false;
 		}
 	}
 
 	@Override
 	public boolean updateFormation(Formation formation) {
-		if(formation.getIdRespCat()!=0 && formation.getStheme()!=null) {
-			dao.updateFormation(formation);
-			return true;
+		//System.out.println(formation.getIdRespCat() +"    " + formation.getStheme().getIdStheme() + "     " +formation.getCodeForm());
+		if(formation.getIdRespCat()!=0 && formation.getStheme()!=null && formation.getCodeForm()!=null) {
+			return dao.updateFormation(formation);
 		}else {
-			System.err.println("Impossible d'ajouter cette formation car ID RespCat ou ID Theme est pas valide");
+			System.err.println("Impossible de modifier cette formation d'Id : " + formation.getIdFormation());
 			return false;
 		}
 	}
 
 	@Override
 	public boolean deleteFormation(int id) {
-		if(id>0) {
-			dao.deleteFormation(id);
-			return true;
+		if(id>0 && idExist(id)) {
+			return dao.deleteFormation(id);
+			//return true;
 		}else {
 			return false;
 		}
 	}
 
 	@Override
-	public List<Formation> searchByParams(String codeForm, String nomForm, String nomSthem) {
-		return dao.searchByParams(codeForm, nomForm, nomSthem);
+	public List<Formation> searchByParams(String codeForm, String nomForm, String nomStheme) {
+		if(codeForm != null && nomForm != null && nomStheme !=null) {
+			return dao.searchByParams(codeForm, nomForm, nomStheme);
+		}else {
+			return null;
+		}
 	}
